chore(admin-cli): compliment `Dispatch` with `Run` + restructure accordingly
**Note that this only refactors the `credential` command module to start, making it a little easier to digest and look at what a command module will look like as proposed. If we go ahead with it, I'll do the others in subsequent PR/PRs.**

So we don't have the commit history of this anymore, but I had previously introduced the `Dispatch` trait to give every top-level command module (e.g. `credential`, `machine`, `sku`, etc.) a unified interface for how commands get dispatched from `main.rs` (as well as any nested subcommand groups). This structured `main.rs` and the entire crate itself, by giving each command it's own module with a `mod.rs` dispatcher.

With that in place, we then ended up with a lot of boilerplate command handling with `Cmd::` structs, passing through context (possibly de-structuring args), and then calling corresponding `cmd::` functions. Soooo, I thought I could take the structure of the `Dispatch` handler for command groups, and push it down even further, by introducing a `Run` handler alongside the "leaf" commands.

The idea is these exist as impls on the `Cmd::` variant's argument struct, giving the `impl` direct access to the args (as `self`), and then executing the underlying command handler directly. This gives us a two layer `Dispatch` + `Run` model for how we execute commands through the CLI. For nested command groups, you could of course even do `Dispatch` + `Dispatch` + `Run`.

Adding a new subcommand now becomes a pattern of:
1. Define your `Args` struct.
2. Add the variant to your `Cmd` enum.
3. Implement `Run` directly on the `Args` struct.
4. Update your `.dispatch()` to `.run()` your leaf command.

You will have your `main.rs` with a bunch of:

```
match command {
    CliCommand::Credential(cmd) => cmd.dispatch(ctx).await?,
    .....
}
```

And then your command-module `mod.rs` with a bunch of:

```
impl Dispatch for Cmd {
    async fn dispatch(self, mut ctx: RuntimeContext) -> CarbideCliResult<()> {
        match self {
            Cmd::AddUFM(args) => args.run(&mut ctx).await,
            Cmd::DeleteUFM(args) => args.run(&mut ctx).await,
            .....
        }
    }
}
```

And then an `args.rs` with `Run` impls right along each argument definition:

```
impl Run for AddUFMCredential {
    async fn run(self, ctx: &mut RuntimeContext) -> CarbideCliResult<()> {
        // your command handling logic here -- making API calls, etc.
    }
}
```

With this layout, the struct definition, `clap` attributes, and execution logic are all in one place. If you want to know what `add-ufm` does, you look in one file.

NOW, ANOTHER thing that happened while I was doing this was: I had these `Run` impls in `args.rs`, but they were just calling the command handling logic from `cmds.rs`. I realized I could just fold the code directly into the `Run` impl and get rid of `cmds.rs` entirely. BUT, that would mean we'd lose any sense of structuring whatsoever, which isn't nice, so THEN I realized I could break out each leaf command into its own module, with its own mod-level `Run` handler, simple `args.rs`, and `cmd.rs`.

Now, instead of doing a PR by PR thing, I'm just dropping a complete restructuring. It's a large refactor, but I think it provides both file structure and code structure with the implementations of `Dispatch`, `Run`, and modules per command.

The layout is:
1. `main.rs` pulls in top-level commands and calls their `Dispatch` handler for command.
2. Each command pulls in it's sub-commands and calls either: a. `.dispatch()` via another nested command group `Dispatch` handler. b. `.run()` for a leaf command `Run` handler.
3. The leaf command module is: a. Named the same as the command itself, making it easy to align the structure of the command with the filesystem layout. b. Where both the args struct and code live, making it easy to go to one place to look at everything.

It's worth mentioning that some leaf commands share the same underlying argument struct. For example, `boot-override get` and `boot-override clear` both take the same `BootOverride` args. Since we can now only implement `Run` once per args type, we use a newtype pattern to wrap the common args type to give it a leaf command specific type, e.g.

```
// GetBootOverride wraps the shared BootOverride as a subcommand
// specific newtype to allow sharing of BootOverride, and still
// providing a subcommand-specific Run trait implementation.
pub struct GetBootOverride {
    #[clap(flatten)]
    pub inner: BootOverride,
}

impl Run for GetBootOverride {
    async fn run(self, ctx: &mut RuntimeContext) -> CarbideCliResult<()> {
        cmds::get(self.inner, &ctx.api_client).await
    }
}
```

`#[clap(flatten)]` makes the newtype transparent to `clap` (same CLI args), and `.inner` gives direct access to the wrapped type. The only thing that changes is each subcommand now has its own type to define a `Run` impl on.
