+++ crates/admin-cli/src/credential/add_bmc/cmd.rs
@@ -14,8 +14,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-pub mod cli_options;
-pub mod dispatch;
-pub mod measurement;
-pub mod run;
-pub mod runtime;
+
+use ::rpc::admin_cli::CarbideCliResult;
+use ::rpc::{CredentialType, forge as forgerpc};
+
+use super::args::Args;
+use crate::credential::common::password_validator;
+use crate::rpc::ApiClient;
+
+pub async fn add_bmc(c: Args, api_client: &ApiClient) -> CarbideCliResult<()> {
+    let password = password_validator(c.password)?;
+    let req = forgerpc::CredentialCreationRequest {
+        credential_type: CredentialType::from(c.kind).into(),
+        username: c.username,
+        password,
+        mac_address: c.mac_address.map(|mac| mac.to_string()),
+        vendor: None,
+    };
+    api_client.0.create_credential(req).await?;
+    Ok(())
+}
