+++ crates/admin-cli/src/credential/add_ufm/args.rs
@@ -14,8 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-pub mod cli_options;
-pub mod dispatch;
-pub mod measurement;
-pub mod run;
-pub mod runtime;
+
+use clap::Parser;
+
+#[derive(Parser, Debug, Clone)]
+pub struct Args {
+    #[clap(long, required(true), help = "The UFM url")]
+    pub url: String,
+
+    #[clap(long, default_value(""), help = "The UFM token")]
+    pub token: String,
+}
