+++ crates/admin-cli/src/credential/add_ufm/cmd.rs
@@ -14,8 +14,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-pub mod cli_options;
-pub mod dispatch;
-pub mod measurement;
-pub mod run;
-pub mod runtime;
+
+use ::rpc::admin_cli::CarbideCliResult;
+use ::rpc::{CredentialType, forge as forgerpc};
+
+use super::args::Args;
+use crate::credential::common::url_validator;
+use crate::rpc::ApiClient;
+
+pub async fn add_ufm(c: Args, api_client: &ApiClient) -> CarbideCliResult<()> {
+    let username = url_validator(c.url)?;
+    let password = c.token;
+    let req = forgerpc::CredentialCreationRequest {
+        credential_type: CredentialType::Ufm.into(),
+        username: Some(username),
+        password,
+        mac_address: None,
+        vendor: None,
+    };
+    api_client.0.create_credential(req).await?;
+    Ok(())
+}
