+++ crates/admin-cli/src/credential/mod.rs
@@ -15,35 +15,68 @@
  * limitations under the License.
  */
 
-pub mod args;
-pub mod cmds;
+mod add_bmc;
+mod add_dpu_factory_default;
+mod add_host_factory_default;
+mod add_nmxm;
+mod add_uefi;
+mod add_ufm;
+mod common;
+mod delete_bmc;
+mod delete_nmxm;
+mod delete_ufm;
+mod generate_ufm_cert;
 
 #[cfg(test)]
 mod tests;
 
 use ::rpc::admin_cli::CarbideCliResult;
-pub use args::Cmd;
+use clap::Parser;
 
 use crate::cfg::dispatch::Dispatch;
+use crate::cfg::run::Run;
 use crate::cfg::runtime::RuntimeContext;
 
+#[derive(Parser, Debug, Clone)]
+#[clap(rename_all = "kebab_case")]
+pub enum Cmd {
+    #[clap(about = "Add UFM credential")]
+    AddUFM(add_ufm::Args),
+    #[clap(about = "Delete UFM credential")]
+    DeleteUFM(delete_ufm::Args),
+    #[clap(about = "Generate UFM credential")]
+    GenerateUFMCert(generate_ufm_cert::Args),
+    #[clap(about = "Add BMC credentials")]
+    AddBMC(add_bmc::Args),
+    #[clap(about = "Delete BMC credentials")]
+    DeleteBMC(delete_bmc::Args),
+    #[clap(
+        about = "Add site-wide DPU UEFI default credential (NOTE: this parameter can be set only once)"
+    )]
+    AddUefi(add_uefi::Args),
+    #[clap(about = "Add manufacturer factory default BMC user/pass for a given vendor")]
+    AddHostFactoryDefault(add_host_factory_default::Args),
+    #[clap(about = "Add manufacturer factory default BMC user/pass for the DPUs")]
+    AddDpuFactoryDefault(add_dpu_factory_default::Args),
+    #[clap(about = "Add NmxM credentials")]
+    AddNmxM(add_nmxm::Args),
+    #[clap(about = "Delete NmxM credentials")]
+    DeleteNmxM(delete_nmxm::Args),
+}
+
 impl Dispatch for Cmd {
-    async fn dispatch(self, ctx: RuntimeContext) -> CarbideCliResult<()> {
+    async fn dispatch(self, mut ctx: RuntimeContext) -> CarbideCliResult<()> {
         match self {
-            Cmd::AddUFM(args) => cmds::add_ufm(args, &ctx.api_client).await,
-            Cmd::DeleteUFM(args) => cmds::delete_ufm(args, &ctx.api_client).await,
-            Cmd::GenerateUFMCert(args) => cmds::generate_ufm_cert(args, &ctx.api_client).await,
-            Cmd::AddBMC(args) => cmds::add_bmc(args, &ctx.api_client).await,
-            Cmd::DeleteBMC(args) => cmds::delete_bmc(args, &ctx.api_client).await,
-            Cmd::AddUefi(args) => cmds::add_uefi(args, &ctx.api_client).await,
-            Cmd::AddHostFactoryDefault(args) => {
-                cmds::add_host_factory_default(args, &ctx.api_client).await
-            }
-            Cmd::AddDpuFactoryDefault(args) => {
-                cmds::add_dpu_factory_default(args, &ctx.api_client).await
-            }
-            Cmd::AddNmxM(args) => cmds::add_nmxm(args, &ctx.api_client).await,
-            Cmd::DeleteNmxM(args) => cmds::delete_nmxm(args, &ctx.api_client).await,
+            Cmd::AddUFM(args) => args.run(&mut ctx).await,
+            Cmd::DeleteUFM(args) => args.run(&mut ctx).await,
+            Cmd::GenerateUFMCert(args) => args.run(&mut ctx).await,
+            Cmd::AddBMC(args) => args.run(&mut ctx).await,
+            Cmd::DeleteBMC(args) => args.run(&mut ctx).await,
+            Cmd::AddUefi(args) => args.run(&mut ctx).await,
+            Cmd::AddHostFactoryDefault(args) => args.run(&mut ctx).await,
+            Cmd::AddDpuFactoryDefault(args) => args.run(&mut ctx).await,
+            Cmd::AddNmxM(args) => args.run(&mut ctx).await,
+            Cmd::DeleteNmxM(args) => args.run(&mut ctx).await,
         }
     }
 }
